backend/x11: refactor, prepare support for multiple outputs
